refactor(client): for headers; layout
