+++ client/src/modules/SetEditing.tsx
@@ -12,9 +12,10 @@ import { CardInterface, SetInterface, SetInterfaceDraft } from 'interfaces';
 import style from 'styles/pages/createset.module.scss';
 import { Toggle } from 'ui/Toggle';
 import { AnimatePresence, motion } from 'framer-motion';
-import { transition } from 'ui/Header'; // @todo - fix this
+import { transition } from 'ui/Layout'; // @todo - fix this
 import { isBackendLess } from 'utils/staticData';
 import { useLocalStore } from 'storage/useLocalStore';
+import Header from 'ui/Header';
 
 const SetEditing: NextPage<{ setFigure?: SetInterfaceDraft }> = ({ setFigure }) => {
   const router = useRouter();
@@ -70,20 +71,18 @@ const SetEditing: NextPage<{ setFigure?: SetInterfaceDraft }> = ({ setFigure }) 
 
   return (
     <>
-      <header className={style.header}>
-        <div className={style.header__inner}>
-          <div>
-            <Button onClick={() => router.push(setFigure?.id ? `/${setFigure.id}` : '/')} title="close" variant="icon">
-              <svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24" fill="currentColor">
-                <path d="M0 0h24v24H0V0z" fill="none" />
-                <path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12 19 6.41z" />
-              </svg>
-            </Button>
-            <p>{setFigure && setFigure.id ? 'Update study set' : 'Create study set'}</p>
-          </div>
+      {isImportShown ? null : (
+        <Header>
+          <Button onClick={() => router.push(setFigure?.id ? `/${setFigure.id}` : '/')} title="close" variant="icon">
+            <svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24" fill="currentColor">
+              <path d="M0 0h24v24H0V0z" fill="none" />
+              <path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12 19 6.41z" />
+            </svg>
+          </Button>
+          <p>{setFigure && setFigure.id ? 'Update study set' : 'Create study set'}</p>
           <button onClick={handleSubmit(onSubmit)}>Save</button>
-        </div>
-      </header>
+        </Header>
+      )}
       <div
         className={`container ${style.class}`}
         style={isImportShown ? { overflow: 'hidden', maxHeight: 'calc(100vh - 50px)' } : undefined}
@@ -204,7 +203,7 @@ const Import: React.FC<ImportProps> = ({ setIsImportShown, insertImport }) => {
       cards.map(({ term, definition, ...rest }) => ({
         ...rest,
         term: term?.replace(/^\s+|\s+$/g, '').replace(/\s+/g, ' ') || '',
-        definition: definition?.replace(/^\s+|\s+$/g, '').replace(/\s+/g, ' ') || '',
+        definition: definition?.replace(/^\s+|\s+$/g, '').replace(/\s+/g, ' ') || '', // @todo - refactor, used here twice
       })),
     );
     setIsImportShown(false);
@@ -212,12 +211,10 @@ const Import: React.FC<ImportProps> = ({ setIsImportShown, insertImport }) => {
 
   return (
     <motion.div className={style.import} variants={transition} initial="init" animate="anim" exit="init">
-      <header className={style.header}>
-        <div className={style.header__inner}>
-          <button onClick={() => setIsImportShown(false)}>Cancel</button>
-          <button onClick={onSave}>Import</button>
-        </div>
-      </header>
+      <Header>
+        <button onClick={() => setIsImportShown(false)}>Cancel</button>
+        <button onClick={onSave}>Import</button>
+      </Header>
       <div className={style.import__content}>
         <Input
           value={formState.input}
