+++ client/src/pages/[set]/write.tsx
@@ -12,6 +12,7 @@ import Link from 'next/link';
 import { isAnswerCorrect } from 'utils/isAnswerCorrect';
 import { useLocalStore } from 'storage/useLocalStore';
 import { isBackendLess } from 'utils/staticData';
+import Header from 'ui/Header';
 
 type Results = {
   round: number;
@@ -92,19 +93,18 @@ const WritePage: NextPage<{ pagekey: string }> = ({ pagekey }) => {
   if (!currSet) return <Custom404 />;
   return (
     <>
-      <header className={style.header}>
-        <div className={style.header__inner}>
-          <button onClick={() => push(`/${pagekey}`)} title="close">
-            <svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24" fill="currentColor">
-              <path d="M0 0h24v24H0V0z" fill="none" />
-              <path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12 19 6.41z" />
-            </svg>
-          </button>
-          <Link href="/">
-            <a className={style.header__logo}>Project MWB</a>
-          </Link>
-        </div>
-      </header>
+      <Header>
+        <button onClick={() => push(`/${pagekey}`)} title="close">
+          <svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24" fill="currentColor">
+            <path d="M0 0h24v24H0V0z" fill="none" />
+            <path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12 19 6.41z" />
+          </svg>
+        </button>
+        <Link href="/">
+          <a className={style.header__logo}>Project MWB</a>
+        </Link>
+        <div></div>
+      </Header>
       {cards.length && status !== 'E' ? (
         <div className={style.form}>
           <div className={style.form__inner}>
