+++ client/src/pages/u/[user]/settings.tsx
@@ -1,16 +1,17 @@
 import React from 'react';
 import { NextPage } from 'next';
+import { useRouter } from 'next/router';
 import { useForm } from 'react-hook-form';
 import { useMutation, useQuery } from 'react-query';
 import { authApi } from 'apis/authApi';
 import { UserInterface } from 'interfaces';
 import Custom404 from 'pages/404';
-import style from 'styles/pages/user.module.scss';
 import { notify } from 'utils/notify';
 import { Input } from 'ui/Input';
 import { Button } from 'ui/Button';
-import { useRouter } from 'next/router';
 import { useUserStore } from 'storage/useUserStore';
+import style from 'styles/pages/user.module.scss';
+import Header from 'ui/Header';
 
 const SettingsPage: NextPage = () => {
   const router = useRouter();
@@ -38,20 +39,16 @@ const SettingsPage: NextPage = () => {
   if (!user.data) return <Custom404 />;
   return (
     <>
-      <header className={style.header}>
-        <div className={style.header__inner}>
-          <div>
-            <Button onClick={() => router.push(`/u/${user.data.id}`)} title="close" variant="icon">
-              <svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24" fill="currentColor">
-                <path d="M0 0h24v24H0V0z" fill="none" />
-                <path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12 19 6.41z" />
-              </svg>
-            </Button>
-            <p>Account Settings</p>
-          </div>
-          <button onClick={handleSubmit(onSubmit)}>Save</button>
-        </div>
-      </header>
+      <Header>
+        <Button onClick={() => router.push(`/u/${user.data.id}`)} title="close" variant="icon">
+          <svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24" fill="currentColor">
+            <path d="M0 0h24v24H0V0z" fill="none" />
+            <path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12 19 6.41z" />
+          </svg>
+        </Button>
+        <p>Account Settings</p>
+        <button onClick={handleSubmit(onSubmit)}>Save</button>
+      </Header>
       <div className="container">
         <form className={style.settings}>
           <Input label="Email" {...register('email')} required />
