+++ client/src/styles/components/input.module.scss
@@ -22,6 +22,7 @@
     display: block;
     width: 100%;
     height: 100%;
+    padding: 0.1rem 0;
     outline: none;
   }
 }
