+++ client/src/styles/components/layout.module.scss
@@ -1,3 +1,127 @@
+// top bar
+
+.plus {
+  position: relative;
+
+  &__button {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    svg {
+      margin: 0 -2.5px;
+    }
+  }
+
+  &__menu {
+    position: absolute;
+    top: 110%;
+    right: 10px;
+    z-index: 100;
+    width: 160px;
+    padding-top: 4px;
+    padding-bottom: 4px;
+    margin-top: 2px;
+    border: 1px solid var(--color-outline);
+    border-radius: 10px;
+    color: var(--color-on-surface);
+    background-color: var(--color-surface);
+    box-shadow: var(--box-shadow);
+
+    li {
+      display: block;
+      margin: 0.25rem 0.5rem;
+
+      button,
+      a {
+        display: block;
+        width: 100%;
+        padding: 0.5rem;
+        border-radius: 10px;
+        border: 1px solid transparent;
+        text-align: left;
+        outline: none;
+
+        &:hover,
+        &:focus {
+          color: var(--color-on-primary);
+          background-color: var(--color-primary);
+          border-color: var(--color-outline);
+        }
+      }
+    }
+  }
+}
+
+.avatarbutton {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.5rem;
+  padding: 0.5rem;
+  background-color: var(--color-primary);
+  color: var(--color-on-primary);
+  border-radius: 10px;
+}
+
+.avatarpic {
+  max-height: 40px;
+  max-width: 40px;
+  height: 100%;
+  width: 100%;
+  object-fit: cover;
+  border-radius: 100%;
+}
+
+.menu {
+  z-index: 5;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  color: var(--color-on-background);
+  background-color: var(--color-background);
+
+  &__inner {
+    max-width: var(--container-width);
+    width: 100%;
+    margin: 0 auto;
+  }
+
+  &__list {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+    padding: 1rem;
+
+    li {
+      color: var(--color-on-primary);
+      background-color: var(--color-surface);
+      border-radius: 10px;
+      border: 1px solid var(--color-outline);
+      border-color: var(--color-outline);
+
+      &:hover {
+        background-color: var(--color-primary);
+      }
+
+      a,
+      button {
+        display: flex;
+        align-items: center;
+        gap: 1rem;
+        width: 100%;
+        padding: 0.5rem;
+        border-radius: 10px;
+        text-align: left;
+      }
+    }
+  }
+}
+
+// bottom bar
+
 .navigation {
   position: fixed;
   bottom: 0;
