+++ client/src/styles/globals.scss
@@ -76,6 +76,7 @@ button {
   color: inherit;
   border: 0;
   cursor: pointer;
+  -webkit-tap-highlight-color: #0000;
 
   &:disabled {
     cursor: default;
