+++ client/src/styles/pages/createset.module.scss
@@ -1,30 +1,3 @@
-.header {
-  position: relative;
-  width: 100%;
-  height: 50px;
-
-  color: var(--color-on-surface);
-  background-color: var(--color-surface);
-  box-shadow: var(--box-shadow);
-
-  &__inner {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 1rem;
-    max-width: var(--container-width);
-    height: 100%;
-    margin: 0 auto;
-    padding: 1rem;
-
-    > div {
-      display: flex;
-      align-items: center;
-      gap: 0.5rem;
-    }
-  }
-}
-
 .class {
   &__general {
     display: flex;
