+++ client/src/styles/pages/exam.module.scss
@@ -1,37 +1,3 @@
-// @todo - to refactor header component (used same in flashcards)
-.header {
-  z-index: 5;
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  width: 100%;
-  height: 50px;
-
-  color: var(--color-on-surface);
-  background-color: var(--color-surface);
-  box-shadow: var(--box-shadow);
-
-  &__inner {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    max-width: var(--container-width);
-    height: 100%;
-    margin: 0 auto;
-    padding: 1rem;
-  }
-
-  &__logo {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    text-transform: uppercase;
-    font-weight: 700;
-  }
-}
-
 .container {
   position: relative;
   display: flex;
