+++ client/src/styles/pages/flashcards.module.scss
@@ -1,32 +1,3 @@
-.header {
-  position: relative;
-  width: 100%;
-  height: 50px;
-
-  color: var(--color-on-surface);
-  background-color: var(--color-surface);
-  box-shadow: var(--box-shadow);
-
-  &__inner {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    max-width: var(--container-width);
-    height: 100%;
-    margin: 0 auto;
-    padding: 1rem;
-  }
-
-  &__logo {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    text-transform: uppercase;
-    font-weight: 700;
-  }
-}
-
 .flashcards {
   display: flex;
   flex-direction: column;
